Move state from WritingSession into App
